Implemented CRUD operations for web interface.
CRUD operations have been implemented for all tables, but not all of
them are working correctly; many of the PUT and DELETE calls return
success but the changes are not reflected in future GETs. Cannot tell if
problem is on interface end or API end, need to look into it. User
management is also implemented, though may need to reprogram to use
Azure Tables DB instead of SQL DB. Some of the API calls are now
returning internal server errors after some of my POSTs, may need to
rebuild DB and debug step-by-step to see where the problem is.